some fix

+++ src/main/java/com/example/hotelbooking/repository/RolesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends CrudRepository<Role, Long> {
+public interface RolesRepository extends CrudRepository<Role, Integer> {
 }
